compiler nitpicks, DESCRIPTION update, and excising some under-tested features

+++ src/dbarts/functions.cpp
@@ -176,7 +176,7 @@ namespace dbarts {
     
     for (uint32_t j = 0; j < numValues; ++j) {
       if (values[j] == true) {
-        if (positiveValueCount == i) return (int32_t) j;
+        if (positiveValueCount == i) return static_cast<int32_t>(j);
         ++positiveValueCount;
       }
     }
@@ -202,7 +202,7 @@ namespace dbarts {
     bool rightFound = false;
     
     *leftIndex = 0; // left value if you top out
-    *rightIndex = fit.scratch.numCutsPerVariable[variableIndex] - 1; // right value if you top out
+    *rightIndex = static_cast<int32_t>(fit.scratch.numCutsPerVariable[variableIndex]) - 1; // right value if you top out
     
     bool isRightChild;
